refactor(estimation): consolidate state override preparation logic
